bug #PSBM-67157 aka prlctl disconnect/connect problems
dispatcher config is not updated at the moment a device
added/removed event is consumed. the client sees inconsistent state
though the device state has changed already.
let's place the promise into the vm queue. thus its execution
occurs immediatelly after the configuration update. in case of a
device hotplug we doesn't know its alias beforehand because it is
libvirt generated. let's wait for any device event and pray it is
the one we are waiting for.