bug #PSBM-68676 aka PRL_ERR_BACKUP_RESTORE_PROHIBIT_WHEN_ATTACHED
disabling of a backup device is implemented as a dettached task
triggered by an event about the device disconnect from libvirt.
it is undetermined when the execution starts after the event
emission and the worst we cannot predict when the execution finishes.
the test saw an example of the unfinished execution.
a device unplug is gaurantied to finish after the libvirt device
disconnect event consumption. thus it is safe to call the backup
device disabling logic synchronously right after the call without
complicated deferring of an event handler.