bug #PSBM-68634 aka backup hdd state=disconnected
a backup device is disconnected from the libvirt pov until an attach
is issued explicitly. thus we have a small time window when the
device is disconnected in dispatcher config of the vm. we write that
state on receipt of a defined event after the define in edit vm task.
applyVmConfig code reads config from the cache. its version may
contain that disconnected state. the hotplug code doesn't attach
disconnected devices. the config is not updated. the device stays
disconnected.
we use the cache to read devices aliases required for a graceful
disconnect of devices. let's read them directly from libvirt. this
way we avoid the race.