bug #PSBM-24425 aka Auto-detected VM templates storage on pstorage
we have a brand new managed entity - a shared template. by design
we provide 4 operations on them one of which is listing.
shared templates are supposed to be registered inside a virtual vm
directory for each eligible mount point. these directories should
not persist along with ordinary directories. this patch suggests
to store these directories in a dedicated place and strip from
the catalog dto in the process of storing into a file.
listing of managed entities is suggested to implement using the
chain of responsibility pattern where each chain is responsible to
handle directories of a certain kind. we have a chain for vms,
another for cts and one for shared templates.