bug #PSBM-70128 aka VMs don't force stop after 300 seconds
before the kill by timeout during a stop of a vn we query state of
the vm from libvirt. libvirt doesn't have suspended state and reports
stopped instead. we don't kill the vm then because the stopped state
is an exception and wait for a confirmation of the state transition.
unfortunatelly there won't be an event. we never release the client
job.
let's kill unexceptionally in the timeout handler.