bug #PSBM-71223 aka prlctl report wrong hdd index on create
we have a code which complements a vm configuration obtained from
libvirt with dispatcher specific data. it tries to guess a device
index matching devices from libvirt and dispatcher versions of the
vm config. for hdd it compares system names only to make a guess.
test used some name for a device and libvirt translated it into
an alias. our match doesn't work now because strings are different.
we need something immutable for a fallback comparison.
let's employ serial number if it is not empty in the dispatcher
version of the config.