Get security level on DispToDisp connections #PSBM-63986
GetBackupTree task was not inheriting security level, hence it didnt
work with a High security level set. Looks like this is a global issue,
that needs to be addressed in DispToDisp connection helper. Low security
is represented by 0 hence we do not know for sure whether low level was
passed or it was not set at all. Just in case, if 0 is received, reset
security level with the level obtained from dispatcher minimal security
level configuration.