bug #PSBM-71335 aka prl_disp_service killed by SIGSEGV
null pointer dereference. we were unable to parse input xml. this
is a topic for another issue actually. but still we need to handle
the error path. now we have some error paths. a client is aware
of the output in case an error is detected inside the capabilities
transponster. let's and another condition to turn on the error
path branch.