bug #PSBM-71432 aka prl_nettool not executed after snapshot switch
we have a mech to prevent 2 guest agent detectors to handle a
successful connection to the agent. the mech prevents a network
configuration in guest because another detector has started on
receipt of an agent connected event.
need to have a mark that a detector is in progress to prevent
concurrent execution of detectors. let's place it directly into the
started state and sweep on receipt of an tools state event. we
generate that event in the detector both on success and failure.