bug #PSBM-69333 aka VM uses 'harddisk.hdd.pstorage-target' as HDD
we have a contention between 2 hosts for access to the vm config.
unfortunately, it was the migration target which lose and was unable
to commit patched vm configuration after the hdd merge. the contention
was a result of constant get config requests at the source. also, both
the source and the target were processing the vm events almost
simultaneously and processing involved updating the config.
we should add a special handling for a stop event during migration.
it is better not touch the config then. btw, we grab an exclusive
lock for the operation at the source. the event handling edits grab
another one. we could take advantage of a conflict between these
operations and prevent the unnecessary edits from accessing the
config.
this would block patch of the vm config on stop of the vm vnc
infrastructure when the source vm stops. the tunnel lives in the
main application thread. the conflict will block the thread which
is unacceptable. let's move the handler into the thread pool.