bug #PSBM-72732 aka Unexpected rule is found
we process a running -> paused state transition of the migrated vm.
this involves updating of some cluster related data in the vm
config. but we agreed to not touch the vm config at the source
during migration because of conflicts with the target in the
shared use case. thus the update hangs on attempt to register
an exclusive operation. then it fails anyway but the hang lasts for
2 minutes and we don't make a progress. the following handling of
a paused -> stopped transtion is delayed. it begins when the vm is
deleted at the source. we don't have config then thus we don't
have any source of the firewall rules to disable.
let's skip the cluster action for a migration. in case of success
we don't need to make updates at all. in case of a failure the
state of the vm gets back to running then it is still consistent.