start VNC server for a running VM in the run of a migration
looks like novel libvirt sends an extra state event for a VM during
migration at the target. the VM first is paused and then moves
to running. we expect that the VM was running before turn to paused
thus all the actions associated with transition to running have
already been executed. but they were not.
let's have a special state for a vm which appears paused at the first
onset. then we have a dedicated actions sequence for moving from
that state into running and then we'll start the VNC server.