bug #PSBM-73618 aka Final VE backup size much bigger than all previous backups collectively in case of using "--keep-chain" option
den@ suggests to use the qemu rebase command for backup chain
coalescing.
ok. we need to craft an ephemeral vm for backup images, start it
in paused more, perform the coalescing the same way we do commit
during shared migration and then stop and destroy the vm. the
coalesced images are for removal in the end. the recipient backup
metadata should be updated accordingly.