bug #PSBM-76358 aka backup attach failed due to PRL_ERR_VM_APPLY_CONFIG_FAILED
type-o. backing store is the first alternative in the backing store
choice. without proper initialization it cannot be serialized
properly. new initialization for images is inside the Hdd type directly.
but backup disk is not an image. we need to call setBackingStore
from the base of the Hdd type which has appropriate code.