bug #PSBM-72443 aka libvirt XML error: target 'vda' duplicated for disk sources
a disk target name in libvirt configuration of a vm is constructed
from the disk interface type and stack index value. we choose a
single type for all the devices of the vm. thus stack indexes should
be unique to avoid a names collision. we don't update a device
stack index if m_hddType is PMS_VIRTIO_BLOCK_DEVICE. thus leads to
duplicates when hdd devices of different types are configured for
the vm.
use a transparent assignment of stack indexes based of the selected
hdd type.