bug #PSBM-77285 aka VM missed from 'prlctl list' after upgrade
libvirt hicced up during upgrade. dispatcher performs many long
rpc's inside a libvirt connection handler. another connection
handler may start when the handler is in progress. that new handler
updates the registry resulting in errorneous unregistering of
the registry items. these that cannot be registered due to the
already exists error.
the proposed solution is to prepare lists of items requiring an rpc
beforehand in a connection handler, then update the registry and
perform the rpcs. this way we either won't list vms at all because
of a disconnect or perform a local modification fast avoiding rpc
delays. this should reducing possibility of a clash between 2
connection handlers.