bug #PSBM-77944 aka backup_restore_pcs6 ERROR: prl_disp_service killed by SIGSEGV
we process a finish package at the source side when the task thread
has finished. the handler accesses the task thread context when it
is not there.
we have a synch mech already inside the task implementation which
allows to avoid the problem. unfortunatelly, it is executed when the
task thread disappears. let's move the call to finalyzeTask callback
which runs in the very end of a task execution.