bug #PSBM-78476 aka Migration of running linked clone leaves hdd.pstorage-target
a block commit call without a base value performs commit of the whole
backing chain. it doesn't work in case of a linked clone because
the base snapshot is shared between several vms and is locked in
general case if vms live on a shared storage.
fortunatelly, the block commit can operate in the shallow mode to
merge only the top delta. use it to prevent the problem from the
issue.