bug #PSBM-78644 aka prlctl migrate --clone removes nameserver and gateway from resulting VM config
we match dispatcher and libvirt devices by host interface name. it
changes when we call reset config in the very begining of a migration.
then the match between the reset config and the one which comes from
the source libvirt side fails because the reset rewrites the host
interface names.
a migration in clone mode is allowed for stopped vms only. we have
an extra define for these vms after the libvirt part. let's make
a reset there for clones.