bug #PSBM-78995 aka After dispatcher restart VM is not marked as template if PGVLF_GET_ONLY_IDENTITY_INFO is specified
isTemplate flag from a directory item config is not persisted in vz7.
PGVLF_GET_ONLY_IDENTITY_INFO handler constructs a config template
for a directory item and copies value of the isTemplate flag from
the directory item config. it is always false when the config is
loaded from disk.
let's have a dedicated list handler for both ordinary and ephemeral
templates which sets the flag value.