bug #PSBM-78699 aka "prl-disp-service" crash
we have several use cases when a channel is getting closed while
its tracked devices have already been destroyed. the channel calls
close for each device reference. some of the references are dangling
at the moment. the solution was to prevent a deadlock between the
channel close and the device close which calls back the channel
implicitly.
we have cid generator which tracks channel id assignments. it is
a monitor thus it is safe to call it from multiple threads. the
generator produces unique ids thus removal of a value from the
tracker is successful only once. therefore it is reasonable to
enter the channel critical section for a device reference removal
purpose only when the id assignment removal is successful. then
we won't call a device close from inside channel close when the
device is destroyed because the device destructor has already
removed its id assignment.