bug #PSBM-77405 aka testCdRomDeviceConnect FAIL: cdrom0 isn't connected
applying config on receipt of a cdrom tray event races with a begin
edit call. the client receives a config version which conflicts with
the one after the application.
we don't wait for event processing after a cdrom device update. this
is the reason of the race. let's generate a config updated event in
the end of a tray event processing and send reply when the one is
delivered to the request handler. btw, libvirt doesn't send an event
when a device state is not changed. let's immitate the event handling
by a 10 seconds timeout handler.