bug #PSBM-78370 aka after qemu update VM started without websocket
we initiate the vm config update when both the keep alive channel gets
closed and the socat process terminates. we receive disconnect first
during a qemu process upgrade then initiate the socat process termination
which takes some time and postpones reconfiguration even if the
tunnel is not operational at the moment.
we have 2 stages of a tunnel termination: it is torn down and dismanteled.
then let's emit another signal to indicate that a tunnel enters the
first stage. the reconfiguration occurs when that signal is delivered.