bug #PSBM-80339 aka "prl-disp-service" crash: "prl_disp_service killed by SIGSEGV"
by default invokeMethod uses the auto connection despatch of an invokation.
i.e. the call may be performed synchroneously if the callee is inside
the object thread. this may lead to out of order invokation fi when
one thread calls add timer and then another call from the libvirt
thread accesses the new timer. it is possible because the timer id
is assigned by the add call.
let's switch hub related invokeMethod's into the queued mode to avoid
out of order invokations.