bug #PSBM-80443 aka prl_disp_service killed by SIGSEGV
a saveConfig accesses vm config watcher. the later is reset during
the service object termination. vnc servers teminate on a
qcoreapplication aboutToQuit signal. the dsp service is bogus at
the moment. the config watcher object is unavailable. it is an
error to access it.
drop the config watcher usages from config manager because we don't
need that functionality.