Gracefully shutdown VMs on host reboot #PSBM-79509
Current implementation suspends VMs which are properly configured
for suspend in Autostop. Other VMs must be gracefully stoppd, but
instead we ignored them and systemd just killed the processes
leading to unclean shutdown in such VMs. This patch introduces
clean shutdown for VMs which are not going to be suspended.