skip timestamp tests inside the in memory vm configs cache
generic implementation uses that test to detect external changes
inside a config file. there are no files in case of an in memory
cache. the test produces a false config not found use case.
let's switch to a pure in memory cache and delegate detection of
expired entries to its direct users.