use chipset type/version to read/set machine type of a vm
vms of different machine types are incompatible from the guest pov.
thus we need to choose a type for a new vm and to preserve it
during different stages of the vm lifecycle.
we have the Chipset section in xml model for vms. now it stores
default values. it is suggested to use these values for all the
existing vms.
new vms chipset attributes will be extracted from the machine type
value or assigned explicitly. it is the client obligation to check
compatibility of different chipsets.