Improve output for create_bridges script PSBM-83174
Currently script does not report interfaces which are skipped, and
neither does state the reason. Although script was introduced for
use within anaconda, it is occasionally launched manually by end
users. Since we analyze interfaces and decide whether it needs
a bridge or not, we should also print the decision result instead
of a silent skip. This should allow users to handle misconfig
on their own.

Signed-off-by: Pavel Vokhmyanin <pvokhmyanin@virtuozzo.com>