bug #PVA-37345 aka VA creates VM without SerialNumber on VZ7
VA creates a VM without devices first and then edits the VM config.
we have a loop inside the edit which determines if an extra lock
needs to be acquired. the loop also performs some mutations of devices
configs it scans. if the loop quits before all the devices was
being inspected the mutations may be skipped. this is the case
for VA.
let's scan all the devices and then grab the lock.