bug #PSBM-83854 aka full backup created instead of incremental
QUrl adds // after either the file schema or when a hostname+port
is not empty. both are false for the ploop schema. ploop_open_dev
doesn't accept an url without // after schema and fails. we cannot
read bitmap from a ploop image.
revert to the string concatenation approach in construction of an
image url.