bug #PSBM-83672 aka Driver and guest Tools verification failed, see the errors above in the log
we have a backup of a windows vm restored over a linux one. in the
end the task applies update config to libvirt. unfortunately, it
uses a method which purpose is to update configuration. this is
wrong because we drop the vm history here.
need to overwrite the vm config after a restore. use a proper call.