bug #PSBM-85553 aka Cannot access storage file '/vzt/vzt_785955ab/tmp/tmptgdEBO/tmp1eVs6L/7f0029cf-ce5c-4d19-86f9-befdf81ccf09/har ...
testers keep employ vms with serial ports in their use cases. after
a successful move of hdds libvirt updates its config. we proceed
to removal of the old vm private and fail because serial port items
are busy by the qemu process.
let's fail when a vm is configured with serial ports or uses a efi
bios.