bug #PSBM-85756 aka Cannot unset ve as template
we have a code which updates a vm libvirt config on an new vm event.
it is required to patch the vm config when the vm is being created
directly in libvirt. the code uses a wrong condition to select the
update config branch because it is true both when an unexpected and
an expected  vm is detected. the problem from the issue resulted from
the second save of the vm config on receipt of a defined event
generated by that wrong config update on insert.
we don't set the SourceVmUuid attribute for unexpected vms. let's
use that mark for selection of the update config branch.