bug #PSBM-85876 aka [rhel_to_vz] Incorrect state of Migrated VMs
we use a common registrar for both shared templates and unknown vms.
it sets the template flag to true unconditionally for a new directory
item. thus a vm is treated as a template by the code which calculates
the vm state during processing of a get vm state request. thus the
status value is always stopped.
the template flag is stored inside vm config as well. the config is
a primary source of information when an entry is being registered
inside the directory. let's use the value from config to set
the flag inside the directory item.