ignore started events from libvirt
according to the sources they issue a resumed event and then a
started one. we issue a switch to running event in both cases.
thus we may turn the vm into a wrong state if another state
change event arrives in the middle.
let's skip started events at all. add a log instead just in case.