bug #PSBM-86509 aka N6Legacy4Task15MigrateVmTargetE is not terminated when 20Task_MigrateVmSource fails
due to some unknown reason we don't cancel migration if a connection
to the source is lost during convertion i.e. we don't even try
to do it. that is why the operation doesn't fail in the end because
the convertion completes successfully.
it is better to do cancel and later check if the operation has been
cancelled actually when the process results are inspected.