bug #PSBM-86825 aka Failed to start the VM: Operation timeout. Exit code: 255
due to some unknown reason we skip vm state restoring events
on revert to a snapshot. we use an alternative aproach to update
the vm state. the m_last value of the vm model is not updated.
it stays running when the vm reverts to stopped fi.
let's process state events and rely on a new logic to ignore
duplicate state switches.