bug #PSBM-86857 aka qemu_advanced_suite.checkWholeNodeUpdate ERROR: KeyError: 'websocket'
after an update has finished socat doesn't quit immediately. there
is a time window when the process is still alive and it holds the
port. after the update we issue a restart of socats. they fail to
bind to the port and quit.
vnc related infrastructure for a vm doesn't have any connection to
other vm related stuff. thus it is impossible to know if a tunnel
is active.
let's store links to both rfb and websocket tunnels inside vm
state machine instance and track the tunnel activity. any incomming
request to start a new instance is deferred and lauched only when
the active tunnel deactivates.