cancel a backup get tree target side if its source disconnects
we have a built-in timeout for a inter-dispatcher request-response
round trip. the source task of the pair fails and quits if the
timeout fires but the target side keeps on and continues to
consume resources.
let's spread isConnected checks throught the target side task code
to detect when a cancelation is required. then quit with the
cancelled error.