bug #PSBM-88452 aka  Failed to migrate the VM: Operation failed. Failed to execute the operation.
we try send as many packages as the sender is capable to handle.
rate libvirt traffic just has no chance to get sent. but that
traffic contains important keep alive messages.
let's revert to the original send 1 package and wait for a confirmation
approach to allow libvirt traffic to pass to the other side.