embellish the network listing code
we need to reduce the number of libvirt calls on reading of
networks list. the main problem is getting a list of bridges.
all the bridge related operations pass by reading a list bridges.
let's split bridge listing into front and backend parts and
reuse the later for the network listing problem too.