bug #PSBM-80024 aka CTID is replaced by UUID during restore
vzctl2_env_open_conf is unable to determine the VE ctid if a
caller uses NULL for the ctid argument. thus we always end up with
uuid for registering the VE after a restore.
fortunately, we store a dispatcher version of the VE config inside
a backup. let's try it first and fallback to the vzctl2_env_open_conf
service only if loading of the dispatcher xml fails.