bug #PSBM-89839 aka Dispatcher is unable to configure route for VM sometimes
network manager downs an interface when the one is being removed
from bridge. this is the case when we setup a routed interface.
we chose that approach for routed interfaces design because
libvirt was unable to create a vm interface without a bridge. now
it is able to create an ethernet interface which is what we need
exactly for the routed interface feature.
let's use that type of iterfaces to configure a host routed one
in libvirt config for a vm.