bug #PSBM-89841 aka lin.vm_backup_restore ERROR: EnvExecutionError: eval failed Exit code: 1
the loadFromFile flag means the oposite inside the implementation i.e.
devices paths get converted into absolute ones.
we store these paths being relative in respect to the vm home inside
the vm config thus it is wrong to perform the convertion.
set the flag to false to prevent it.