bug #PSBM-89866 aka Dispatcher gets SIGTERM every day during backup
we use QProcess for starting processes. it provides a callback
waitForStarted which fires when the child definetly starts. otherwise
it returns an error after timeout expiration. then we try to
shutdown the child with a term signal. if the child didn't manage
to exec at the moment of receiving the signal a signal handler
of the dispatcher process got invoked. the signal handler performs
a write to the same channel as the handler in the genuine dispatcher
process does in case of a signal receipt. dispatcher reads an
event from the channel and treats it as a sign of a termination
request then quits.
let's store a pid of the process which registeres sigaction and make
sure that we process signal in context of a process with the same
id. otherwise do nothing.