bug #PSBM-91434 aka Impossible to start VM with attached USB device
by default a usb device is configured with mandatory startup policy
and prevent the vm from starting when the device is unavailable.
fortunatelly, there is optional policy which is what we need to
start a vm when an assigned usb hostdev is disconnected from the
host. let's employ it the same way we did for hdd's