bug #PSBM-90833 ana #PSBM-91944 aka Pass valid chars for devices user aliases
libvirt has an alternative format for user defined aliases of vm
devices. dot character is unacceptable there. a usb device sys path
can contain dots when the device is connected to a hub. libvirt
silently discards the alias and doesn't report about an unsupported
xml format. we expect everything ok and depend on the alias value
for the device disconnect, the later hangs because libvirt generates
another alias and use it in an event acknoledging the disconnect.
fix alias construction for usb's and pull modern usb controller
types supported by latest version of libvirt.