bug #PSBM-90998 aka "prl-disp-service" crash: "prl_disp_service killed by SIGSEGV"
we start a timer in one thread and then move the target object into
another. qt forbids it and a timer event is generated for the
object which has already been destroyed. processing of the event
corrupts the heap.
let's perform an honest starting of a timer inside another qt thread.