bug #PSBM-94237 aka Failed to update the VM: The virtual machine is currently locked by another sessi ...
event has 3 modes: timeout, loose and success. in case of the
timeout we report an error immediately which doesn't allow us to
wait for a full 1 minute timeout. thus we don't wait expected
amount of time for a pending operation to complete.
let's use a dedicated error code for the timeout case and spin
inside the conflict resolution loop to wait for the exclusive lock
again.