refactor conflict resolution
we don't care about return code from the event wait inside a conflict
resolve call. they all are treated as an instruction to keep trying
to resolve the conflict.
let's ignore the status from event wait, log if it fails and quit
with success in case waiting is the choice for the conflict
resolution.