bug #PSBM-96000 aka ate50_iter2 ERROR: prl_disp_service killed by SIGSEGV
it is a libvirt channel exit which we wait for completion to quit from
source pump. we expect that qemu channels are down at the moment
which was always the case before. but this time a socket finished
event stuck somewhere in the queue and was delivered to the slot
when the containing pump was over. we destroyed channel objects thus
their references dangled and we've got ub.
we use the same launch_type type for both source and target pumps
at the target we don't have shared pointers and use naked ones
instead. thus it's impossible to transfer socket instances ownership
using reference counting. anyway we store socket instances reliably
inside a listener list which gets purged on exit from the state.
let's postpone emptying the list to the moment of the state machine
destruction.