bug #PSBM-96083 aka "prl_disp_service killed by SIGSEGV" (no call trace!)
an exec task can be cancelled. the cancel occurs from a thread other
than the task one. boost::optional is not thread-safe actually which
may result in something weird.
let's replace optional with atomic int and use -1 as a sentinel
value because xexec uses that constant to indicate an error.