bug #PSBM-96427 aka "prl-disp-service" crash: "prl_disp_service killed by SIGSEGV"
async job keeps on progress in a pool thread when libvirt host
thread is over.
now we have a dedicated libvirt related pool, we make sure that
all its workers are finished before leaving the host thread. but
still it is better to quit using raw pointers in favour of weak
ones to prevent undesirable cases when a pointee has already
been destroyed but still referenced by a running code.