bug #PSBM-95845 aka "prl-disp-service" crash: "prl_disp_service killed by SIGSEGV"
the problem occurs when a backup task fails. records about thaw
tasks failed because of an internal libvirt lock are in logs. we
have a code which registers a timer in one thread to postpone
thaw execution and an object which is in charge of that execution
which is being destroyed in the backup task thread. the object
uses a mutex to protect its state. the mutex performs writes into
memory. the design assumes that destruction of that object disconnects
a slot safely from the timer. looks like it is not true and the
timer may use that object at the same time it is being destroyed
by the backup thread. the usage then leads to corruption of the
process memory.
let's use a custom destructor of a smart ptr pointee to move
destruction of the thaw executor into its native thread and prevent
concurrency.