bug #PSBM-96764 aka VM template created on pre update 11 node can not be converted to VM on update 11
we have a booking storage for vm registrations. all known vms are
moved to that storage for a later update info after a connection
to libvirt gets established. now we skip ct entries only thus
vm templates permeate into the storage. then a branch for a user
vm registration is selected when an event from libvirt for the
new vm is being processed because the template is in the booking
storage. we cannot match a user by it directory id because there
is no workspace for built-in templates directory. it's impossible
to proceed thus we don't emit an event about the new vm and the
edit task which started defining of the vm in libvirt hangs.
skip vm templates also in the code which decides which objects
require an update from libvirt on reconnect.