improve thaw routine during backup
we try to thaw the vm guest during backup either after a first
protocol package or in the end of nbd stage or from a timeout
handler running in the main thread. the procedure may take much
time and it blocks the main thread if executed after the timeout.
this prevents other activities from execution in that thread which
results in a poor performance.
let's delegate execution to a global pool thread instead and have
a lightweight code to run in the main thread.