workaround duplicate hostnames issue
a user assigned the same hostname to both source and target of a
backup. this was a surprise to the code which checks locality of
the operation and it guessed the backup is local which resulted
in failure of the operation with invalid arg error.
let's use remote peer address of the connection between participants
of a backup operation for the source hostname and fall back to
original approach from command in case the value is empty.