switch a vm being deleted into a transitional state during the task execution.
we need to distinguish when a vm disapears because of a dispatcher
task execution or because libvirt decides so. we send an event
in the first case and need to do the same in the second too. we
cannot just send the event because it would be a duplicate in the
dispatcher task case.
let's switch the vm into a special state from the task explicitly
and send nothing after transition to the final state, i.e. vm
disapears.