Do not install /128 ipv6 route for routed interface #PSBM-99432
We install additional route for ipv6 routed interface in order
to simulate proxy arp behaviour #PSBM-44285

With this route we forward on-link network traffic to next hop
interface which is reachable by link-local address.
Then Host is able to route packets to/from other destinations.

But iwith /128 there is no other on-link network addresses
and the route redirect the ipv6 interface address itself,
which is wrong.

We should not install /128 prefix. Host address and other ipv6
connectivity will work via default route.