bug #OVZ-7132 aka prl_disp_service sends host wide kill(-1, 15) signal under unknown conditions
we have a race when main exec task thread finishes waiting for the
ve process and cancel thread making its progress. the later can
see m_pid > 0 in the beginning, write a log record and then end
with m_pid -1 because the main thread updated the value after the
wait finish.
let's keep the pid to kill in a local variable to avoid the member
updates in the background.