PSBM-103849 Fix remote restore from PVA - refactor local restoration logic.
Local restoration optmization (via file) for ploop Containers was introduced in
commit c09fb05f. But that logic didn't take into account scheme of PVA
backups/restore - when the channel between dispatchers is proxied via local PVA
agents on the nodes - and thus, RestoreSource party considered it local. As a
result, the restore was tried to be made via 'file://' scheme even for remote
restoration.

Move locality determination to RestoreTarget task - because only this party
knows for sure is it local or remote backup.