PSBM-131558
The q35 chipset does not work with ide devices. We decided to change the default interface device type from ide to sata in libprlsdk. This created the problem of duplicate devices. For example sda for CD and sda for hard drive. There is a valid change from sata sd device to hd as we need it to be unique to sata devices.