PSBM-138122 Dispatcher Cannot find created bridge connection
In the process of creating a virtual network, the dispatcher removes ":" from the mac address for its config,
so libvirt has a ":" in the mac address in the network interface, but there is no ":" in the mac address in the dispatcher config.
Thus, we cannot find an already existing network interface.