Add cloudlinux and cloundlinux8 into dist_map.#PSBM-142803
At the moment, cloudlinux and cloundlinux8 are not in dist_map,
which means that it is possible to specify the creation of a dist
VM via prlctl, but the OS type will not be passed to libvirt
and then libvirt will change the dispatcher config.
This all leads to the fact that the OS line is empty in the VM config.
Adding the necessary dist solves this problem.

Signed-off-by: Efim Shevrin <efim.shevrin@virtuozzo.com>