CT backup: do not create useless tmp folder #PSBM-143405 #PSBM-143174
Commented part of old code which creates tmp folder for CT backups.
This folder is unused at this time but can be required in future, so
it is better to comment code for legacy. This folder stays in the
filesystem in cases then prl-disp-service finished work after crash,
reboot or any other unexpected cases.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>