disk: fill in SizeInBytes fields for VmHardDisk #PSBM-143636
SizeInBytes field used to be ignored in many places for disk configuration
and instead of it we used another one getSize() with round to megabytes.
Further it started to lead to problems with creating backups and now we
are gonna fill this parameter correctly and use the exact number of bytes
without round.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>