Expand dispatcher blkio physical_block_size and logical_block_size #PSBM-131181
We add a hint for the user in the form of default values for the logical block and the physical block.
Selecting specific values offering a good balance between disk performance and the use of space in primary memory,
also these parameters allow you to maintain compatibility with older systems.

Signed-off-by: Efim Shevrin <efim.shevrin@virtuozzo.com>