nvram: skip checking nvram with new name
Do not do anything if nvram has proper new VZ name. Even if it
contains old format. Dispatcher does not print any messages in this case
overwise we get a lot of useless messages in evry VM reboot or Dispatcher
restart. 'NVRAM_4M.dat' is a default name for new nvram in Virtuozzo.
If customer change the name of old nvram manually it should be his
responsibility.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>