Fix: chipset upd should take into account already set Os2 params #PSBM-145761
Reverse.cpp: Mixer/Builder::setResources

The Os2 structure is responsible for the Os section in the vm xml config.
Before the chipset parameters were set, the loader and nvram sections
could already be set. The current fix takes into account the loader and
nvram sections and does not change them.

Signed-off-by: Efim Shevrin <efim.shevrin@virtuozzo.com>