vm upgrade: upgrade NVRAM of running VM on reboot #PSBM-144917
Dispatcher should switch on_reboot action and stop VM for upgrading
NVRAM and switching from 2Mb to 4Mb OVMF.
After upgrading NVRAM VM should be started if it was rebooted or still be
stopped if it was halted according to customer request.
Lifecycle Vm parameters should be restored too in customer state.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>