EditVM: fix for misconfiguration between Disp and Libvirt #PSBM-146887
During development of Dispatcher found bug about misconfiguration
between libvirt and Dispatcher. This error should not happen in current
version of Dispatcher, but who knows. For this purposes we have implemented
current mechanism - 'prlctl set VM --update-nvram'.
But it does not work because dispatcher sees that NVRAM was updated.
So, here we just only do not upgrade NVRAM, we reply to prlctl
success message and prlct starts 'set_updated' functionality
and updates domain.xml in libvirt.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>