TaskMoveVm:In case of error Vm dir will be saved on target #PSBM-147476
If an error occurs during the execution of TaskMoveVm at some stage,
for example, when deleting the source directory,then the directory
with the VM will remain on the target and it could be registered
on taget by prlctl. It is also important to say that in case of an
error, the config of the source VM will not be changed,
in particular, the home directory of the VM will not change, it will
remain the same as before TaskMoveVm exec.

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>