restart: upgrade of VM should restore last state #PSBM-148420
In cases 'prlctl restart VM' the dispatcher calls 'virDomainReboot' function,
which works not as we expected and does not send a Reboot event to
the Dispatcher.
'virDomainReboot' checks the lifecycle action 'on_reboot' and powers off
the VM in cases of 'destroy' value and sends only events about shutdown VM.
This patch emulates sending 'reboot' event from libvirt and calls proper
handling methods for further upgrade.

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>