Change mechanism for start and stop bridge netif. #PSBM-148660
Mechanism for collecting data about network adapters uses
data from /sys/class/net, which was added as part of f5d6d2a0bfa.
Thus, the bridge interface inside the dispatcher is constructed
using the interface name from /sys/class/net instead of the
virInterfacePtr pointer from the libvirt, while this pointer
remains NULL in our case. The consequence of the null pointer
is that we cannot use the libvirt api to start or stop the interface,
namely virInterfaceCreate and virInterfaceDestroy.
We replace the &virInterfaceCreate and virInterfaceDestroy
calls with ifup and ifdown calls, respectively.

Signed-off-by: Fima Shevrin <efim.shevrin@virtuozzo.com>