transponster: all appended CD devices should have boot order #PSBM-150310
When VM boots on UEFI BIOS (edk2-ovmf) and has a cdrom attached via IDE bus,
it may not see the device. In particular, this blocks us from installing
the guest tools or upgrade OS.
So let's add boot order attribute every time we attach CDROM to a VM.
This patch affects the running VM (live migration and installtools).

Signed-off-by: Aleksandr Leskin <aleksandr.leskin@virtuozzo.com>